+++ src/main/kotlin/com/terraformation/backend/search/field/GramsField.kt
@@ -17,7 +17,7 @@ class GramsField(
     override val fieldName: String,
     override val displayName: String,
     override val databaseField: TableField<*, BigDecimal?>,
-    override val table: SearchTable
+    override val table: SearchTable,
 ) : SingleColumnSearchField<BigDecimal>() {
   override val supportedFilterTypes: Set<SearchFilterType> =
       EnumSet.of(SearchFilterType.Exact, SearchFilterType.Range)
