+++ src/main/kotlin/com/terraformation/backend/search/table/SpeciesTable.kt
@@ -1,0 +1,29 @@
+package com.terraformation.backend.search.table
+
+import com.terraformation.backend.db.FuzzySearchOperators
+import com.terraformation.backend.db.tables.references.ACCESSIONS
+import com.terraformation.backend.db.tables.references.SPECIES
+import com.terraformation.backend.search.SearchTable
+import com.terraformation.backend.search.SublistField
+import com.terraformation.backend.search.field.SearchField
+import org.jooq.Record
+import org.jooq.TableField
+
+class SpeciesTable(tables: SearchTables, fuzzySearchOperators: FuzzySearchOperators) :
+    SearchTable(fuzzySearchOperators) {
+  override val primaryKey: TableField<out Record, out Any?>
+    get() = SPECIES.ID
+
+  override val sublists: List<SublistField> by lazy {
+    with(tables) {
+      listOf(
+          accessions.asMultiValueSublist("accessions", SPECIES.ID.eq(ACCESSIONS.SPECIES_ID)),
+      )
+    }
+  }
+
+  override val fields: List<SearchField> =
+      listOf(
+          textField("name", "Species name", SPECIES.NAME),
+      )
+}
